resolve conflicts with develop
